feat(RELEASE-1191): diferentiate pyxis and umb in signing part1 (#734)
This commit modifies the `simple-signing-pipeline` and its tasks to use
separate credentials for Pyxis and UMB authentication. Prior to this
commit, both were pulled from the same configMap values so it was
impossible to use stage for one system and prod for the other.

Because e2e uses the `simple-signing-pipeline` on the stage sre cluster
vs a git resolver (for now), the `request-and-upload-signature` changes
can't be part of this commit. If they were, the
`simple-signing-pipeline` used in this E2E run would fail as it is the
previous version which does not pass the new
`request-and-upload-signature` parameters.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>