refactor(RELEASE-1319): move pipelines into managed/internal dirs (#749)
* refactor(RELEASE-1319): move pipelines into managed/internal dirs

This commit moves all pipeline definitions in the pipelines/ directory
into pipelines/managed or pipelines/internal. It also creates sym links
to the previous location for all the pipelines that were migrated. This
is for backwards compatibility; they will be removed once all
ReleasePlanAdmissions and other references to the pipelines are
migrated.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>