refactor(RELEASE-1319): move tasks based on type (#763)
This commit moves all the tasks in this directory. Most of the tasks are
now in tasks/managed. This is for all tasks used directly in a managed
pipeline. There is also tasks/internal for tasks used in
internal-request or internal-pipelinerun pipelines and tasks/collectors
for collector pipeline tasks.

The commit also updates the CI for the tekton unit tests for the new
task locations.

A sym link was created for the internal tasks that were in `tasks` to
their previous locations. This is because the `simple-signing-pipeline`
which calls them is not always called via git resolver, so it is
temporary backwards compatibility.

Finally, the commit modifies the definition of each pipeline as the
pathInRepo value of each task git resolver has to be updated for the new
location. This comes with a version bump.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>